restructure and move functions
