+++ packages/react/src/components/CombinationChart.tsx
@@ -12,7 +12,8 @@ import { UpSetSizeInfo } from '../derive/deriveSizeDependent';
 import { UpSetStyleInfo } from '../derive/deriveStyleDependent';
 import { UpSetSelection } from './interfaces';
 import UpSetDot from './UpSetDot';
-import { addonPositionGenerator, clsx } from './utils';
+import { addonPositionGenerator } from './utils';
+import { clsx } from '../utils';
 
 const CombinationChart = React.memo(function CombinationChart<T>({
   d,
