+++ packages/react/src/venn/components/VennChart.tsx
@@ -9,7 +9,8 @@ import { ISetLike } from '@upsetjs/model';
 import React, { PropsWithChildren } from 'react';
 import { VennDiagramDataInfo } from '../derive/deriveVennDataDependent';
 import { VennDiagramStyleInfo } from '../derive/deriveVennStyleDependent';
-import { clsx, wrap } from '../../components/utils';
+import { wrap } from '../../components/utils';
+import { clsx } from '../../utils';
 import VennArcSlice from './VennArcSlice';
 import VennCircle from './VennCircle';
 import VennUniverse from './VennUniverse';
